recognition/arcface_paddle/requirement.txt
visualdl
opencv-python
pillow
numpy
easydict
scipy
sklearn
requests
prettytable
tqdm
Pillow
scikit-learn==0.23.2
opencv-python==4.4.0.46
onnxruntime
onnx
paddle2onnx
